+++ CHANGELOG.md
@@ -1,12 +1,23 @@
 # Changelog
 
-## [0.5.0] - 2025-12
+## [0.5.0] - 2025-12-10
 
 ### Added
- - add required `task_type` attribute to OpenquakeHazardSolution (breaking change).
+ - add required `task_type` attribute to `OpenquakeHazardSolution` (breaking change).
+ - new `RuptureSet` subclasses File with specific attributes.
+ - new ENV variables:
+    - `DB_READ_ONLY` for safe testing agains LIVE backends, default = False.
+    - `CLOUDWATCH_ENABLED` for disabling cloud metrics, default = True.
+    - `MIGRATE_FILE_TO_RUPTSET` for `RuptureSet` class migrations, default = False.
+ - new resolvers for FileData objects:`post_url_v2` and `post_data_v2`
+    to make upload testing possible, and to better align with boto3 API. 
+    NB `nshm-toshi-client`should migrate to these, but not required immediately.
+    
+### Removed
+ - deprecated mutation operations for `RuptureGenerationTask`. Use `AutomationTask` instead.
 
 ### Changed
- - make `task_type` attribute on OpenquakeHazardTask `required` (breaking change).
+ - for `AutomationTask` the mutation task_type = Required is now enforced.
 
 ## [0.4.0] - 2025-12-02